jdk/decode: fix & feat: fix duplicated jvm record & return/throw support
Fix:
Since PT data is splitted, for a single cpu: The splitted data
might result in duplicated record of jvm info: such as Exception,
and Deopt, and dataloss of sideband.
To fix it: We decode it in advance while splitting to get a pair
of start time and end time, so we can ignore events to record before
start time and drain events to record before end time.

Feat:
Instead dump method exit, return bytecode and throw bytecode
is more helpful for decoding, so use return and athrow.